Rename JS function to select_from_weighted
