+++ src/js/mod.rs
@@ -80,11 +80,8 @@ pub fn pick(randomness: &str, n: u32, input: Box<[JsValue]>) -> Result<Box<[JsVa
 //// Picks 1 element from a JavaScript weighted list and returns it.
 #[wasm_bindgen]
 #[allow(dead_code)] // exported via wasm_bindgen
-pub fn pick_one_from_weighted_list(
-    randomness: &str,
-    input: Box<[JsValue]>,
-) -> Result<JsValue, JsValue> {
-    Ok(implementations::pick_one_from_weighted_list_impl(
+pub fn select_from_weighted(randomness: &str, input: Box<[JsValue]>) -> Result<JsValue, JsValue> {
+    Ok(implementations::select_from_weighted_impl(
         randomness, input,
     )?)
 }
@@ -228,7 +225,7 @@ mod implementations {
         Ok(picked.into_boxed_slice())
     }
 
-    pub fn pick_one_from_weighted_list_impl(
+    pub fn select_from_weighted_impl(
         randomness_hex: &str,
         input: Box<[JsValue]>,
     ) -> Result<JsValue, JsError> {
